PipDad
I am a content writer for a family-owned company in Ohio. My wife and I house 14 rescued cats and kittens and try to rescue and re-home all stray cats in our neighborhood. We currently have 8 regular strays, we have heated winter cat houses, and we feed them all twice a day. Our monthly cost just for the 8+ strays is up to $400/month. Our household income is in the $75,000/yr range. When you include the $1,100/month to feed, provide litter and other necessities, vet bills, and all other expenses; we're losing our financial grip. But I've determined that, before I die, we will save as many as we can. They are my legacy, we have no children, and I've convinced my wife of almost 30 years to help the rescue efforts. We are proudly poor because our money goes to help those who cannot help themselves in this human-controlled world. We can't hold back the evil of most people, but we can spread a little love and care for them every day.